Remove references to the parent entry for owners manual
Some entries were published in Contentful that should not have been. Resetting these entries back to draft status seems to break the production build because there are no entries with a parent.

As I understand it, the query is looking for entries with a parent `contentfulparent` and if the published entries don't have a parent assigned, then this query fails.

In the future, we can revert this change, but for now it's the only way I was able to figure out how to fix the production build after un-publishing entries with a parent.