Show time with webinar date, add date-fns for better formatting
The `formatString` method in Gatsby/GraphQL uses moment for formatting and it didn't seem to output the correct time. date-fns was added to provide better time formatting and conversion based on the UTC offset set in Contentful. The time on the webinar page will update to the user's current time zone.